requirements.txt
Pillow==9.4.0
pip==23.0.1
wheel==0.40.0
confine-controller==1.0.2
setuptools==65.5.0
pygame==2.3.0
numpy==1.24.2
six==1.16.0
pynput==1.7.6
scipy==1.10.1
simpleaudio==1.0.4